+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.7
+
+- Add grade and graduationYear fields to MirimUser.
+- Support snake_case and camelCase token responses from the OAuth server.
+
 ## 0.0.1
 
 * TODO: Describe initial release.